requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
duckdb==1.1.3
pandas==2.2.2
pyarrow==17.0.0
azure-storage-blob
python-dotenv
